Move components reusable to ux module. Issue #497
