+++ client/mobrender/redux/action-creators/async-add-block.js
@@ -2,11 +2,11 @@ import * as t from '../action-types'
 import { createAction } from './create-action'
 import AuthSelectors from '~client/account/redux/selectors'
 
-export default ({ mobilizationId, ...block }) => (dispatch, getState, { api }) => {
+export default ({ mobilization_id, ...block }) => (dispatch, getState, { api }) => {
   const credentials = AuthSelectors(getState()).getCredentials()
   dispatch(createAction(t.ADD_BLOCK_REQUEST))
   return api
-    .post(`/mobilizations/${mobilizationId}/blocks`, { block }, { headers: credentials })
+    .post(`/mobilizations/${mobilization_id}/blocks`, { block }, { headers: credentials })
     .then(res => {
       const { widgets_attributes, ...data } = res.data
       dispatch(createAction(t.ADD_BLOCK_SUCCESS, data))
