+++ client/ux/components/index.js
@@ -1,0 +1,4 @@
+export { default as Button } from './button'
+export { default as DivFloat } from './div-float'
+
+export { default as SettingsForm } from './settings-form'
